SonarQube corrections 4

+++ src/main/java/tn/esprit/spring/kaddem/services/ContratServiceImpl.java
@@ -2,81 +2,90 @@ package tn.esprit.spring.kaddem.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import lombok.extern.slf4j.Slf4j;
 import tn.esprit.spring.kaddem.entities.Contrat;
 import tn.esprit.spring.kaddem.entities.Etudiant;
-import tn.esprit.spring.kaddem.entities.Specialite;
 import tn.esprit.spring.kaddem.repositories.ContratRepository;
 import tn.esprit.spring.kaddem.repositories.EtudiantRepository;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
 @Slf4j
 @Service
-public class ContratServiceImpl implements IContratService{
-ContratRepository contratRepository;
+public class ContratServiceImpl implements IContratService {
+	@Autowired
+	ContratRepository contratRepository;
+
+	@Autowired
 	EtudiantRepository etudiantRepository;
-	public List<Contrat> retrieveAllContrats(){
+
+	public List<Contrat> retrieveAllContrats() {
 		return contratRepository.findAll();
 	}
 
-	public Contrat updateContrat (Contrat  ce){
+	public Contrat updateContrat(Contrat ce) {
 		return contratRepository.save(ce);
 	}
 
-	public  Contrat addContrat (Contrat ce){
+	public Contrat addContrat(Contrat ce) {
 		return contratRepository.save(ce);
 	}
 
-	public Contrat retrieveContrat (Integer  idContrat){
+	public Contrat retrieveContrat(Integer idContrat) {
 		return contratRepository.findById(idContrat).orElse(null);
 	}
 
-	public  void removeContrat(Integer idContrat){
-		Contrat c=retrieveContrat(idContrat);
-		contratRepository.delete(c);
+	public void removeContrat(Integer idContrat) {
+		Contrat c = retrieveContrat(idContrat);
+		if (c != null) {
+			contratRepository.delete(c);
+		}
 	}
 
+	public Contrat affectContratToEtudiant(Integer idContrat, String nomE, String prenomE) {
+		Etudiant e = etudiantRepository.findByNomEAndPrenomE(nomE, prenomE);
+		Contrat ce = contratRepository.findByIdContrat(idContrat);
+		Set<Contrat> contrats = e.getContrats();
+		Integer nbContratsActifs = 0;
 
-
-	public Contrat affectContratToEtudiant (Integer idContrat, String nomE, String prenomE){
-		Etudiant e=etudiantRepository.findByNomEAndPrenomE(nomE, prenomE);
-		Contrat ce=contratRepository.findByIdContrat(idContrat);
-		Set<Contrat> contrats= e.getContrats();
-		Integer nbContratssActifs=0;
 		if (!contrats.isEmpty()) {
 			for (Contrat contrat : contrats) {
-				if (((contrat.getArchive())!=null)&& ((contrat.getArchive())))  {
-					nbContratssActifs++;
+				if (contrat.getArchive() != null && contrat.getArchive()) {
+					nbContratsActifs++;
 				}
 			}
 		}
-		if (nbContratssActifs<=4){
-		ce.setEtudiant(e);
-		contratRepository.save(ce);}
+
+		if (nbContratsActifs <= 4) {
+			ce.setEtudiant(e);
+			contratRepository.save(ce);
+		}
 		return ce;
 	}
-	public 	Integer nbContratsValides(Date startDate, Date endDate){
+
+	public Integer nbContratsValides(Date startDate, Date endDate) {
 		return contratRepository.getnbContratsValides(startDate, endDate);
 	}
 
-	public void retrieveAndUpdateStatusContrat(){
-		List<Contrat>contrats=contratRepository.findAll();
-		List<Contrat>contrats15j=null;
-		List<Contrat>contratsAarchiver=null;
+	public void retrieveAndUpdateStatusContrat() {
+		List<Contrat> contrats = contratRepository.findAll();
+		List<Contrat> contrats15j = new ArrayList<>();
+		List<Contrat> contratsAarchiver = new ArrayList<>();
+
 		for (Contrat contrat : contrats) {
 			Date dateSysteme = new Date();
 			if (!contrat.getArchive()) {
-				long difference_In_Time = dateSysteme.getTime() - contrat.getDateFinContrat().getTime();
-				long difference_In_Days = (difference_In_Time / (1000 * 60 * 60 * 24)) % 365;
-				if (difference_In_Days==15){
+				long differenceInTime = dateSysteme.getTime() - contrat.getDateFinContrat().getTime();
+				long differenceInDays = (differenceInTime / (1000 * 60 * 60 * 24)) % 365;
+
+				if (differenceInDays == 15) {
 					contrats15j.add(contrat);
-					log.info(" Contrat : " + contrat);
+					log.info("Contrat: " + contrat);
 				}
-				if (difference_In_Days==0) {
+				if (differenceInDays == 0) {
 					contratsAarchiver.add(contrat);
 					contrat.setArchive(true);
 					contratRepository.save(contrat);
@@ -84,28 +93,31 @@ ContratRepository contratRepository;
 			}
 		}
 	}
-	public float getChiffreAffaireEntreDeuxDates(Date startDate, Date endDate){
+
+	public float getChiffreAffaireEntreDeuxDates(Date startDate, Date endDate) {
 		long differenceInTime = endDate.getTime() - startDate.getTime();
-		float differenceInDays = (float)(differenceInTime / (1000 * 60 * 60 * 24)) % 365;
-		float differenceInMonths =differenceInDays/30;
-        List<Contrat> contrats=contratRepository.findAll();
-		float chiffreAffaireEntreDeuxDates=0;
+		float differenceInDays = (float) differenceInTime / (1000 * 60 * 60 * 24) % 365;
+		float differenceInMonths = differenceInDays / 30;
+
+		List<Contrat> contrats = contratRepository.findAll();
+		float chiffreAffaireEntreDeuxDates = 0;
+
 		for (Contrat contrat : contrats) {
-			if (contrat.getSpecialite()== Specialite.IA){
-				chiffreAffaireEntreDeuxDates+=(differenceInMonths*300);
-			} else if (contrat.getSpecialite()== Specialite.CLOUD) {
-				chiffreAffaireEntreDeuxDates+=(differenceInMonths*400);
-			}
-			else if (contrat.getSpecialite()== Specialite.RESEAUX) {
-				chiffreAffaireEntreDeuxDates+=(differenceInMonths*350);
-			}
-			else //if (contrat.getSpecialite()== Specialite.SECURITE)
-			 {
-				 chiffreAffaireEntreDeuxDates+=(differenceInMonths*450);
+			switch (contrat.getSpecialite()) {
+				case IA:
+					chiffreAffaireEntreDeuxDates += (differenceInMonths * 300);
+					break;
+				case CLOUD:
+					chiffreAffaireEntreDeuxDates += (differenceInMonths * 400);
+					break;
+				case RESEAUX:
+					chiffreAffaireEntreDeuxDates += (differenceInMonths * 350);
+					break;
+				default:
+					chiffreAffaireEntreDeuxDates += (differenceInMonths * 450);
+					break;
 			}
 		}
 		return chiffreAffaireEntreDeuxDates;
 	}
-
-
 }
